% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_policies.R
\name{get_policies}
\alias{get_policies}
\title{get list of all avaliable policies}
\usage{
get_policies(key)
}
\arguments{
\item{key}{A character. Your API key from rada4you.org}
}
\value{
Dataframe with all available policies
}
\description{
get list of all avaliable policies
}
