% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_policy_details.R
\name{get_policy_details}
\alias{get_policy_details}
\title{Get short description of the policy}
\usage{
get_policy_details(id = 1, key)
}
\arguments{
\item{id}{Vector of policy ids (should consist at least of one value). Default value is '1', which returns test policy}

\item{key}{A character. Your API key from rada4you.org}
}
\value{
Dataframe with 4 variables for each policy: id, name, description, provisional
}
\description{
Get short description of the policy
}
