% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps.R
\name{mps}
\alias{mps}
\title{Отримайте перелік всіх народних депутатів}
\usage{
mps(key)
}
\arguments{
\item{key}{Ваш персональний ключ API, отриманий на rada4you.org}
}
\value{
Датафрейм, який містить всі доступні політик
}
\description{
Отримайте перелік всіх народних депутатів
}
