% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_details.R
\name{policy_details}
\alias{policy_details}
\title{Отримайте короткий опис політики}
\usage{
policy_details(id = 1, key)
}
\arguments{
\item{id}{Вектор, який містить id політик (має складатись щонайменше з одного елементу). За замовчуванням має значення '1', що відповідає тестовій політиці}

\item{key}{Ваш персональний ключ API, отриманий на rada4you.org}
}
\value{
Датафрейм, що складається з чотирьох змінних для кожної політики: id, name, description, provisional
}
\description{
Отримайте короткий опис політики
}
